server/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.1
Flask-Login==0.6.3
python-socketio==5.11.0
eventlet==0.33.3
bcrypt==4.1.3
pyjwt==2.8.0
python-dotenv==1.0.1
opencv-python==4.9.0.80
numpy==1.24.3
mediapipe==0.10.8
pyautogui==0.9.54
pymongo==4.6.3
dnspython==2.6.1
waitress==2.1.2  # Production WSGI server for Windows